Refactor: Move workflow tasks to workflow_tasks
