+++ corehq/motech/openmrs/workflow_tasks.py
@@ -5,15 +5,15 @@ from datetime import timedelta
 
 from corehq.motech.const import DIRECTION_EXPORT
 from corehq.motech.exceptions import ConfigurationError
-from corehq.motech.openmrs.const import PERSON_UUID_IDENTIFIER_TYPE_ID
+from corehq.motech.openmrs.const import (
+    ADDRESS_PROPERTIES,
+    NAME_PROPERTIES,
+    PERSON_PROPERTIES,
+    PERSON_UUID_IDENTIFIER_TYPE_ID,
+)
 from corehq.motech.openmrs.repeater_helpers import (
-    CreatePatientIdentifierTask,
-    CreatePersonAttributeTask,
-    CreateVisitTask,
-    DeletePersonAttributeTask,
-    UpdatePatientIdentifierTask,
-    UpdatePersonAttributeTask,
     get_ancestor_location_openmrs_uuid,
+    get_unknown_encounter_role,
     get_unknown_location_uuid,
 )
 from corehq.motech.openmrs.serializers import to_omrs_datetime
@@ -198,3 +198,464 @@ class CreateVisitsEncountersObsTask(WorkflowTask):
                     )
                 )
         return subtasks
+
+
+class CreatePersonAttributeTask(WorkflowTask):
+
+    def __init__(self, requests, person_uuid, attribute_type_uuid, value):
+        self.requests = requests
+        self.person_uuid = person_uuid
+        self.attribute_type_uuid = attribute_type_uuid
+        self.value = value
+        self.attribute_uuid = None
+
+    def run(self):
+        response = self.requests.post(
+            '/ws/rest/v1/person/{person_uuid}/attribute'.format(person_uuid=self.person_uuid),
+            json={'attributeType': self.attribute_type_uuid, 'value': self.value},
+            raise_for_status=True,
+        )
+        self.attribute_uuid = response.json()['uuid']
+
+    def rollback(self):
+        # if attribute_uuid is not set, it would be because the workflow task to create the attribute failed
+        if self.attribute_uuid:
+            self.requests.delete(
+                '/ws/rest/v1/person/{person_uuid}/attribute/{attribute_uuid}'.format(
+                    person_uuid=self.person_uuid, attribute_uuid=self.attribute_uuid
+                ),
+                raise_for_status=True,
+            )
+
+
+class DeletePersonAttributeTask(WorkflowTask):
+
+    def __init__(self, requests, person_uuid, attribute_uuid, attribute_type_uuid, existing_value):
+        self.requests = requests
+        self.person_uuid = person_uuid
+        self.attribute_uuid = attribute_uuid
+        self.attribute_type_uuid = attribute_type_uuid
+        self.existing_value = existing_value
+
+    def run(self):
+        self.requests.delete(
+            '/ws/rest/v1/person/{person_uuid}/attribute/{attribute_uuid}'.format(
+                person_uuid=self.person_uuid, attribute_uuid=self.attribute_uuid
+            ),
+            raise_for_status=True,
+        )
+
+    def rollback(self):
+        self.requests.post(
+            '/ws/rest/v1/person/{person_uuid}/attribute'.format(person_uuid=self.person_uuid),
+            json={'attributeType': self.attribute_type_uuid, 'value': self.existing_value},
+            raise_for_status=True,
+        )
+
+
+class UpdatePersonAttributeTask(WorkflowTask):
+
+    def __init__(self, requests, person_uuid, attribute_uuid, attribute_type_uuid, value, existing_value):
+        self.requests = requests
+        self.person_uuid = person_uuid
+        self.attribute_uuid = attribute_uuid
+        self.attribute_type_uuid = attribute_type_uuid
+        self.value = value
+        self.existing_value = existing_value
+
+    def run(self):
+        self.requests.post(
+            '/ws/rest/v1/person/{person_uuid}/attribute/{attribute_uuid}'.format(
+                person_uuid=self.person_uuid, attribute_uuid=self.attribute_uuid
+            ),
+            json={
+                'value': self.value,
+                'attributeType': self.attribute_type_uuid,
+            },
+            raise_for_status=True,
+        )
+
+    def rollback(self):
+        self.requests.post(
+            '/ws/rest/v1/person/{person_uuid}/attribute/{attribute_uuid}'.format(
+                person_uuid=self.person_uuid, attribute_uuid=self.attribute_uuid
+            ),
+            json={
+                'value': self.existing_value,
+                'attributeType': self.attribute_type_uuid,
+            },
+            raise_for_status=True,
+        )
+
+
+class CreatePatientIdentifierTask(WorkflowTask):
+
+    def __init__(self, requests, patient_uuid, identifier_type_uuid, identifier):
+        self.requests = requests
+        self.patient_uuid = patient_uuid
+        self.identifier_type_uuid = identifier_type_uuid
+        self.identifier = identifier
+        self.identifier_uuid = None
+
+    def run(self):
+        response = self.requests.post(
+            '/ws/rest/v1/patient/{patient_uuid}/identifier'.format(patient_uuid=self.patient_uuid),
+            json={'identifierType': self.identifier_type_uuid, 'identifier': self.identifier},
+            raise_for_status=True,
+        )
+        self.identifier_uuid = response.json()['uuid']
+
+    def rollback(self):
+        if self.identifier_uuid:
+            self.requests.delete(
+                '/ws/rest/v1/patient/{patient_uuid}/identifier/{identifier_uuid}'.format(
+                    patient_uuid=self.patient_uuid, identifier_uuid=self.identifier_uuid
+                ),
+                raise_for_status=True,
+            )
+
+
+class UpdatePatientIdentifierTask(WorkflowTask):
+
+    def __init__(self, requests, patient_uuid, identifier_uuid, identifier_type_uuid, identifier,
+                 existing_identifier):
+        self.requests = requests
+        self.patient_uuid = patient_uuid
+        self.identifier_uuid = identifier_uuid
+        self.identifier_type_uuid = identifier_type_uuid
+        self.identifier = identifier
+        self.existing_identifier = existing_identifier
+
+    def run(self):
+        self.requests.post(
+            '/ws/rest/v1/patient/{patient_uuid}/identifier/{identifier_uuid}'.format(
+                patient_uuid=self.patient_uuid, identifier_uuid=self.identifier_uuid
+            ),
+            json={
+                'identifier': self.identifier,
+                'identifierType': self.identifier_type_uuid,
+            },
+            raise_for_status=True,
+        )
+
+    def rollback(self):
+        self.requests.post(
+            '/ws/rest/v1/patient/{patient_uuid}/identifier/{identifier_uuid}'.format(
+                patient_uuid=self.patient_uuid, identifier_uuid=self.identifier_uuid
+            ),
+            json={
+                'identifier': self.existing_identifier,
+                'identifierType': self.identifier_type_uuid,
+            },
+            raise_for_status=True,
+        )
+
+
+class CreateVisitTask(WorkflowTask):
+
+    def __init__(self, requests, person_uuid, provider_uuid, start_datetime, stop_datetime, values_for_concept,
+                 encounter_type, openmrs_form, visit_type, location_uuid=None):
+        self.requests = requests
+        self.person_uuid = person_uuid
+        self.provider_uuid = provider_uuid
+        self.start_datetime = start_datetime
+        self.stop_datetime = stop_datetime
+        self.values_for_concept = values_for_concept
+        self.encounter_type = encounter_type
+        self.openmrs_form = openmrs_form
+        self.visit_type = visit_type
+        self.location_uuid = location_uuid
+        self.visit_uuid = None
+
+    def run(self):
+        subtasks = []
+        if self.visit_type:
+            visit = {
+                'patient': self.person_uuid,
+                'visitType': self.visit_type,
+                'startDatetime': self.start_datetime,
+                'stopDatetime': self.stop_datetime,
+            }
+            if self.location_uuid:
+                visit['location'] = self.location_uuid
+            response = self.requests.post('/ws/rest/v1/visit', json=visit, raise_for_status=True)
+            self.visit_uuid = response.json()['uuid']
+
+        subtasks.append(
+            CreateEncounterTask(
+                self.requests, self.person_uuid, self.provider_uuid, self.start_datetime, self.values_for_concept,
+                self.encounter_type, self.openmrs_form, self.visit_uuid, self.location_uuid
+            )
+        )
+        return subtasks
+
+    def rollback(self):
+        if self.visit_uuid:
+            self.requests.delete('/ws/rest/v1/visit/{uuid}'.format(uuid=self.visit_uuid), raise_for_status=True)
+
+
+class CreateEncounterTask(WorkflowTask):
+
+    def __init__(self, requests, person_uuid, provider_uuid, start_datetime, values_for_concept, encounter_type,
+                 openmrs_form, visit_uuid, location_uuid=None):
+        self.requests = requests
+        self.person_uuid = person_uuid
+        self.provider_uuid = provider_uuid
+        self.start_datetime = start_datetime
+        self.values_for_concept = values_for_concept
+        self.encounter_type = encounter_type
+        self.openmrs_form = openmrs_form
+        self.visit_uuid = visit_uuid
+        self.location_uuid = location_uuid
+        self.encounter_uuid = None
+
+    def run(self):
+        subtasks = []
+        encounter = {
+            'encounterDatetime': self.start_datetime,
+            'patient': self.person_uuid,
+            'encounterType': self.encounter_type,
+        }
+        if self.openmrs_form:
+            encounter['form'] = self.openmrs_form
+        if self.visit_uuid:
+            encounter['visit'] = self.visit_uuid
+        if self.location_uuid:
+            encounter['location'] = self.location_uuid
+        if self.provider_uuid:
+            encounter_role = get_unknown_encounter_role(self.requests)
+            encounter['encounterProviders'] = [{
+                'provider': self.provider_uuid,
+                'encounterRole': encounter_role['uuid']
+            }]
+        response = self.requests.post('/ws/rest/v1/encounter', json=encounter, raise_for_status=True)
+        self.encounter_uuid = response.json()['uuid']
+
+        for concept_uuid, values in self.values_for_concept.items():
+            for value in values:
+                subtasks.append(
+                    CreateObsTask(
+                        self.requests, self.encounter_uuid, concept_uuid, self.person_uuid, self.start_datetime,
+                        value, self.location_uuid
+                    )
+                )
+        return subtasks
+
+    def rollback(self):
+        if self.encounter_uuid:
+            self.requests.delete(
+                '/ws/rest/v1/encounter/{uuid}'.format(uuid=self.encounter_uuid), raise_for_status=True
+            )
+
+
+class CreateObsTask(WorkflowTask):
+
+    def __init__(self, requests, encounter_uuid, concept_uuid, person_uuid, start_datetime, value,
+                 location_uuid=None):
+        self.requests = requests
+        self.encounter_uuid = encounter_uuid
+        self.concept_uuid = concept_uuid
+        self.person_uuid = person_uuid
+        self.start_datetime = start_datetime
+        self.value = value
+        self.location_uuid = location_uuid
+        self.obs_uuid = None
+
+    def run(self):
+        observation = {
+            'concept': self.concept_uuid,
+            'person': self.person_uuid,
+            'obsDatetime': self.start_datetime,
+            'encounter': self.encounter_uuid,
+            'value': self.value,
+        }
+        if self.location_uuid:
+            observation['location'] = self.location_uuid
+        response = self.requests.post('/ws/rest/v1/obs', json=observation, raise_for_status=True)
+        self.obs_uuid = response.json()['uuid']
+
+    def rollback(self):
+        if self.obs_uuid:
+            self.requests.delete('/ws/rest/v1/obs/{uuid}'.format(uuid=self.obs_uuid), raise_for_status=True)
+
+
+class UpdatePersonNameTask(WorkflowTask):
+
+    def __init__(self, requests, info, openmrs_config, person):
+        self.requests = requests
+        self.info = info
+        self.openmrs_config = openmrs_config
+        self.person = person
+        self.person_uuid = person['uuid']
+        self.name_uuid = person['preferredName']['uuid']
+
+    def run(self):
+        properties = {
+            property_: value_source.get_value(self.info)
+            for property_, value_source in self.openmrs_config.case_config.person_preferred_name.items()
+            if (
+                property_ in NAME_PROPERTIES and
+                value_source.check_direction(DIRECTION_EXPORT) and
+                value_source.get_value(self.info)
+            )
+        }
+        if properties:
+            self.requests.post(
+                '/ws/rest/v1/person/{person_uuid}/name/{name_uuid}'.format(
+                    person_uuid=self.person_uuid,
+                    name_uuid=self.name_uuid,
+                ),
+                json=properties,
+                raise_for_status=True,
+            )
+
+    def rollback(self):
+        """
+        Reset the name changes back to their original values, which are
+        taken from the patient details that OpenMRS returned at the
+        start of the workflow.
+        """
+        properties = {
+            property_: self.person['preferredName'].get(property_)
+            for property_ in self.openmrs_config.case_config.person_preferred_name.keys()
+            if property_ in NAME_PROPERTIES
+        }
+        if properties:
+            self.requests.post(
+                '/ws/rest/v1/person/{person_uuid}/name/{name_uuid}'.format(
+                    person_uuid=self.person_uuid,
+                    name_uuid=self.name_uuid,
+                ),
+                json=properties,
+                raise_for_status=True,
+            )
+
+
+class CreatePersonAddressTask(WorkflowTask):
+
+    def __init__(self, requests, info, openmrs_config, person):
+        self.requests = requests
+        self.info = info
+        self.openmrs_config = openmrs_config
+        self.person = person
+        self.person_uuid = person['uuid']
+        self.address_uuid = None
+
+    def run(self):
+        properties = {
+            property_: value_source.get_value(self.info)
+            for property_, value_source in self.openmrs_config.case_config.person_preferred_address.items()
+            if (
+                property_ in ADDRESS_PROPERTIES and
+                value_source.check_direction(DIRECTION_EXPORT) and
+                value_source.get_value(self.info)
+            )
+        }
+        if properties:
+            response = self.requests.post(
+                '/ws/rest/v1/person/{person_uuid}/address/'.format(person_uuid=self.person_uuid),
+                json=properties,
+                raise_for_status=True,
+            )
+            self.address_uuid = response.json()['uuid']
+
+    def rollback(self):
+        if self.address_uuid:
+            self.requests.delete(
+                '/ws/rest/v1/person/{person_uuid}/address/{address_uuid}'.format(
+                    person_uuid=self.person_uuid,
+                    address_uuid=self.address_uuid,
+                ),
+                raise_for_status=True,
+            )
+
+
+class UpdatePersonAddressTask(WorkflowTask):
+
+    def __init__(self, requests, info, openmrs_config, person):
+        self.requests = requests
+        self.info = info
+        self.openmrs_config = openmrs_config
+        self.person = person
+        self.person_uuid = person['uuid']
+        self.address_uuid = person['preferredAddress']['uuid']
+
+    def run(self):
+        properties = {
+            property_: value_source.get_value(self.info)
+            for property_, value_source in self.openmrs_config.case_config.person_preferred_address.items()
+            if (
+                property_ in ADDRESS_PROPERTIES and
+                value_source.check_direction(DIRECTION_EXPORT) and
+                value_source.get_value(self.info)
+            )
+        }
+        if properties:
+            self.requests.post(
+                '/ws/rest/v1/person/{person_uuid}/address/{address_uuid}'.format(
+                    person_uuid=self.person_uuid,
+                    address_uuid=self.address_uuid,
+                ),
+                json=properties,
+                raise_for_status=True,
+            )
+
+    def rollback(self):
+        properties = {
+            property_: self.person['preferredAddress'].get(property_)
+            for property_ in self.openmrs_config.case_config.person_preferred_address.keys()
+            if property_ in ADDRESS_PROPERTIES
+        }
+        if properties:
+            self.requests.post(
+                '/ws/rest/v1/person/{person_uuid}/address/{address_uuid}'.format(
+                    person_uuid=self.person_uuid,
+                    address_uuid=self.address_uuid,
+                ),
+                json=properties,
+                raise_for_status=True,
+            )
+
+
+class UpdatePersonPropertiesTask(WorkflowTask):
+
+    def __init__(self, requests, info, openmrs_config, person):
+        self.requests = requests
+        self.info = info
+        self.openmrs_config = openmrs_config
+        self.person = person
+
+    def run(self):
+        properties = {
+            property_: value_source.get_value(self.info)
+            for property_, value_source in self.openmrs_config.case_config.person_properties.items()
+            if (
+                property_ in PERSON_PROPERTIES and
+                value_source.check_direction(DIRECTION_EXPORT) and
+                value_source.get_value(self.info)
+            )
+        }
+        if properties:
+            self.requests.post(
+                '/ws/rest/v1/person/{person_uuid}'.format(person_uuid=self.person['uuid']),
+                json=properties,
+                raise_for_status=True,
+            )
+
+    def rollback(self):
+        """
+        Reset person properties back to their original values, which
+        are taken from the patient details that OpenMRS returned at the
+        start of the workflow.
+        """
+        properties = {
+            property_: self.person.get(property_)
+            for property_ in self.openmrs_config.case_config.person_properties.keys()
+            if property_ in PERSON_PROPERTIES
+        }
+        if properties:
+            self.requests.post(
+                '/ws/rest/v1/person/{person_uuid}'.format(person_uuid=self.person['uuid']),
+                json=properties,
+                raise_for_status=True,
+            )
